Rename file name into types for shared components
